Move types to appropriate files
