+++ common/src/codegen/index.ts
@@ -3,7 +3,7 @@ import { CodeGenMetaFactory } from "./codegenMetaFactory";
 import { createOutputProjectMetadata } from "./codegenOutputProjectMeta";
 import { createSourceProjectMetadata } from "./codegenSourceProjectMeta";
 import { IOutputProjectMetadataProcessor } from "./playwright-charp-common/outputProjectMetadataProcessor";
-import { ICodeGen, SourceFileValidationError } from "./types";
+import { ICodeGen } from "./types";
 
 export {
   CodeGenMetaFactory,
@@ -12,5 +12,4 @@ export {
   generateCode,
   IOutputProjectMetadataProcessor,
   ICodeGen,
-  SourceFileValidationError,
 };
