+++ common/src/file-defs/shared.ts
@@ -1,3 +1,6 @@
+import { ActionType } from "./actionType";
+import { Indent } from "./rmProjFile";
+
 /** Simple dictionary with key of string an value of type T */
 export interface IDictionary<T> {
   [Key: string]: T;
@@ -23,4 +26,33 @@ export interface ITestActionStep extends Omit<ITestStepBase, "type"> {
 export interface ITestStepComment extends Omit<ITestStepBase, "type"> {
   type: "comment";
   comment?: string;
+}
+
+/** Map from `Indent` to actual character of that `Indent` */
+export const indentCharMap = new Map<Indent, string>([
+  [Indent.Tabs, "\t"],
+  [Indent.Spaces, " "],
+]);
+
+export class SourceFileValidationError {
+  /** Name of the file */
+  readonly fileName: string;
+  /** Full path of the file */
+  readonly filePath: string;
+  /** Line number (i.e step index) of the error step */
+  readonly lineNumber: number;
+  /** Error message TEMPLATE, which might contain placeholders for filling with StringRes */
+  readonly message: string;
+  /** Action type */
+  readonly actionType?: ActionType;
+  /**
+   *
+   */
+  constructor(fileName: string, filePath: string, lineNumber: number, message: string, actionType?: ActionType) {
+    this.fileName = fileName;
+    this.filePath = filePath;
+    this.lineNumber = lineNumber;
+    this.message = message;
+    this.actionType = actionType;
+  }
 }
